+++ src/window_data.h
@@ -7,8 +7,8 @@ struct window_data
   window_data();
 
   bool v_sync{false};
-  uint16_t width{0};
-  uint16_t height{0};
+  unsigned int width{0};
+  unsigned int height{0};
   double mouse_x{0.f};
   double mouse_y{0.f};
   float last_mouse_x{0.f};
