+++ src/window_props.cpp
@@ -1,5 +1,5 @@
 
 #include "window_props.h"
 
-window_props::window_props(const std::string &title, uint16_t width, uint16_t height)
+window_props::window_props(const std::string &title, unsigned int width, unsigned int height)
     : title(title), width(width), height(height) {}
